requirements.txt
numpy 1.18.1
pandas 1.0.3
pytorch 1.4.0
pytorch-lightning 0.7.1
scikit-learn 0.22.1
